requirements.txt
amqp==5.2.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.3
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-click==2.3.0
django-debug-toolbar==4.3.0
django-filter==24.1
django-timezone-field==6.1.0
djangorestframework==3.15.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
kombu==5.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
prompt-toolkit==3.0.43
pycparser==2.21
Pygments==2.17.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.3
requests==2.31.0
rich==13.7.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.2
